move inputStyles to ConfigurableObject/Inputs
