+++ DashAI/front/src/components/ConfigurableObject/Inputs/InputStyles.jsx
@@ -1,0 +1,6 @@
+import { TextField } from "@mui/material";
+import { styled } from "@mui/material/styles";
+
+export const Input = styled(TextField)(({ theme }) => ({
+  minWidth: "20vw",
+}));
